+++ generative_app/core/app_pages/chat.py
@@ -1,14 +1,21 @@
 import enum
-import re
 import streamlit as st
 from streamlit.delta_generator import DeltaGenerator
 import chains.llm as llm
 from auth.auth_connection import AuthSingleton
-from templates.template_app import template_app
 from utils.parser import parse_current_app
 import ui.chat_init as chat_init
 from chains.llm import parse
-from ui.end_trial import trial_title, thanks, pay, share, download, download_info, no_code
+from ui.end_trial import (
+    trial_title,
+    thanks,
+    pay,
+    share,
+    download,
+    download_info,
+    no_code,
+)
+
 
 class CommandResult(enum.Enum):
     UNKNOWN = [0, "Unknown command"]
@@ -17,8 +24,9 @@ class CommandResult(enum.Enum):
     RESET = [3, "Code resetted"]
     SAVE = [4, "Code saved"]
 
+
 class ChatBot:
-    def __init__(self, user_id: int, username:str, python_script_path: str):
+    def __init__(self, user_id: int, username: str, python_script_path: str):
         self.python_script_path = python_script_path
         self.background_tasks = set()
         self.user_id = user_id
@@ -26,14 +34,28 @@ class ChatBot:
         self.auth = AuthSingleton().get_instance()
         self.user_role = self.auth.get_user_role(self.user_id)
 
-    def apply_code(self, code:str):
+    def append_code_history(self, code: str):
+        """Store the code into the last_code session data"""
+        if "last_code" not in st.session_state:
+            st.session_state["last_code"] = []
+        st.session_state.last_code.append(code)
+
+    def pop_code_history(self):
+        """Store the code into the last_code session data"""
+        if "last_code" not in st.session_state:
+            return None
+        else:
+            st.session_state.last_code.pop()
+
+    def apply_code(self, code: str):
         if code is None:
             return
         # save code to database
         self.auth.set_code(self.user_id, code)
+        self.append_code_history(code)
 
     @staticmethod
-    def check_commands(instruction:str) -> CommandResult or None:
+    def check_commands(instruction: str) -> CommandResult or None:
         if instruction.startswith("/undo"):
             if "messages" not in st.session_state:
                 return CommandResult.NOTUNDO
@@ -56,7 +78,9 @@ class ChatBot:
         if command == CommandResult.NOTUNDO:
             chat_placeholder.error("Nothing to undo")
         if command == CommandResult.UNDO:
-            self.apply_code(st.session_state.last_code)
+            # remove the last code
+            self.pop_code_history()
+            self.apply_code(st.session_state.last_code[-1])
             # remove the last 4 keys
             if len(st.session_state.messages) > 1:
                 for _ in range(min(len(st.session_state.messages), 3)):
@@ -76,18 +100,26 @@ class ChatBot:
             if code is None or code == "pass\n":
                 chat_placeholder.error("No code to save")
                 return
-            chat_placeholder.info("Download the file by clicking on the button below.\nYou can then run it with `streamlit run streamlit_app.py`")
-            chat_placeholder.download_button(label="Download app", file_name= "streamlit_app.py",
-                                             mime='text/x-python', data=code)
+            chat_placeholder.info(
+                "Download the file by clicking on the button below.\n"
+                "You can then run it with `streamlit run streamlit_app.py`"
+            )
+            chat_placeholder.download_button(
+                label="Download app",
+                file_name="streamlit_app.py",
+                mime="text/x-python",
+                data=code,
+            )
 
     def reset_chat(self):
         print("resetting chat")
         st.session_state["messages"] = {
-            "message_0":
-                {
-                    "role": "assistant",
-                    "content": chat_init.message_en.format(name=self.username) if st.session_state.lang == "en" else chat_init.message_fr.format(name=self.username)
-                },
+            "message_0": {
+                "role": "assistant",
+                "content": chat_init.message_en.format(name=self.username)
+                if st.session_state.lang == "en"
+                else chat_init.message_fr.format(name=self.username),
+            },
         }
         self.save_chat_history_to_database()
         st.session_state.chat_history = []
@@ -95,7 +127,9 @@ class ChatBot:
 
     def add_message(self, role: str, content: str):
         idx = len(st.session_state.messages)
-        st.session_state.messages.update({f"message_{idx}": {"role": role, "content": content}})
+        st.session_state.messages.update(
+            {f"message_{idx}": {"role": role, "content": content}}
+        )
 
     def end_of_trial(self):
         st.title(trial_title[1 if st.session_state.lang == "fr" else 0])
@@ -107,7 +141,12 @@ class ChatBot:
             st.header(download[1 if st.session_state.lang == "fr" else 0])
             c1, c2 = st.columns([1, 0.5])
             c1.markdown(download_info[1 if st.session_state.lang == "fr" else 0])
-            if c2.download_button(label=download[1 if st.session_state.lang == "fr" else 0], file_name= "streamlit_app.py", mime='text/x-python', data=code):
+            if c2.download_button(
+                label=download[1 if st.session_state.lang == "fr" else 0],
+                file_name="streamlit_app.py",
+                mime="text/x-python",
+                data=code,
+            ):
                 st.balloons()
             # TODO add tutorial on how to deploy on web
             # TODO download already compiled app (like .exe)
@@ -115,13 +154,9 @@ class ChatBot:
             st.warning(no_code[1 if st.session_state.lang == "fr" else 0])
 
     def setup(self):
-        # Save last code
-        st.session_state["last_code"] = self.auth.get_code(self.user_id)
-
         if self.user_role == "guest" and self.check_tries_exceeded():
             self.end_of_trial()
             return
-
 
         # If this is the first time the chatbot is launched reset it and the code
         # Add saved messages
@@ -137,6 +172,7 @@ class ChatBot:
                         message = f"```python\n{code}\n```"
                         ex = st.expander("Code")
                         ex.code(code)
+                        self.append_code_history(code)
                     st.markdown(explanation)
                 else:
                     st.markdown(message["content"])
@@ -150,9 +186,9 @@ class ChatBot:
 
     def setup_chat(self):
         # Setup user input
-        if instruction := st.chat_input(f"Tell me what to do, or ask me a question"):
+        if instruction := st.chat_input("Tell me what to do, or ask me a question"):
             if self.user_role == "guest":
-                tries_left = 5 - st.session_state.tries
+                tries_left = st.secrets["tries"] - st.session_state.tries
                 if tries_left <= 0:
                     print("end of trial")
                     st.experimental_rerun()
@@ -172,14 +208,26 @@ class ChatBot:
                 user_message_placeholder.markdown(instruction)
                 with assistant_message_placeholder:
                     current_assistant_message_placeholder = st.empty()
-                    chain = llm.load_conversation_chain(current_assistant_message_placeholder, st.session_state.openai_api_key)
+                    chain = llm.load_conversation_chain(
+                        current_assistant_message_placeholder,
+                        st.session_state.openai_api_key,
+                    )
                     message = ""
 
-                    with st.spinner("⌛Processing... Please keep this page open until the end of my response."):
-                        # Wait for the response of the LLM and display a loading message in the meantime
+                    with st.spinner(
+                        "⌛Processing... Please keep this page open until the end of my response."
+                    ):
+                        # Wait for the response of the LLM and display a
+                        # loading message in the meantime
                         try:
                             print("calling chain")
-                            llm_result = chain({"question": instruction, "chat_history": st.session_state.chat_history, "python_code": st.session_state.last_code})
+                            llm_result = chain(
+                                {
+                                    "question": instruction,
+                                    "chat_history": st.session_state.chat_history,
+                                    "python_code": st.session_state.last_code[-1],
+                                }
+                            )
                         except Exception as e:
                             current_assistant_message_placeholder.error(f"Error...{e}")
                             raise
@@ -198,10 +246,14 @@ class ChatBot:
                     ex.code(code)
                     container.markdown(explanation)
                     if security_rules_offended:
-                        container.warning("Your instruction does not comply with our security measures (code generated will not be populated). See the docs for more information.")
+                        container.warning(
+                            "Your instruction does not comply with our security measures"
+                            "(code generated will not be populated). See the docs for "
+                            "more information."
+                        )
                     if self.user_role == "guest":
                         st.session_state.tries = self.auth.increment_tries(self.user_id)
-                        tries_left = 5 - st.session_state.tries
+                        tries_left = st.secrets["tries"] - st.session_state.tries
                         if tries_left == 0:
                             container.error("You have 0 try left.")
                         elif tries_left == 1:
@@ -214,7 +266,7 @@ class ChatBot:
 
     def check_tries_exceeded(self) -> bool:
         tries = self.auth.get_tries(self.user_id)
-        if tries < 5:
+        if tries < st.secrets["tries"]:
             return False
         return True
 
@@ -226,4 +278,4 @@ class ChatBot:
 
     def save_chat_history_to_database(self):
         print("saving chat history")
-        self.auth.set_message_history(self.user_id,  st.session_state.messages)
+        self.auth.set_message_history(self.user_id, st.session_state.messages)