+++ generative_app/core/app_pages/login_app.py
@@ -1,24 +1,27 @@
 import os
-import sys
 import time
 from typing import Dict, Union
 import shutil
 from pathlib import Path
+
 import streamlit as st
-from auth.auth_connection import AuthSingleton
-import ui.chat_init as chat_init
 from hydralit import HydraHeadApp
 from streamlit.delta_generator import DeltaGenerator
+import streamlit.components.v1 as components
+
+from auth.auth_connection import AuthSingleton
+import ui.chat_init as chat_init
 
 
 class LoginApp(HydraHeadApp):
     """
-    This is an example login application to be used to secure access within a HydraApp streamlit application.
-    This application implementation uses the allow_access session variable and uses the do_redirect method if the login check is successful.
-
+    This is an example login application to be used to secure access
+    within a HydraApp streamlit application.
+    This application implementation uses the allow_access session variable
+    and uses the do_redirect method if the login check is successful.
     """
 
-    def __init__(self, title = '', **kwargs):
+    def __init__(self, title="", **kwargs):
         self.__dict__.update(kwargs)
         self.title = title
         self.auth = AuthSingleton().get_instance()
@@ -38,9 +41,18 @@ class LoginApp(HydraHeadApp):
             print("Auto login detected of user_id: ", user_id)
             self.redirect_after_login(user_id, self.auth.get_username_from_id(user_id))
 
-        st.markdown("<h1 style='text-align: center;'>Login to ChatbotX 💫</h1>", unsafe_allow_html=True)
+        st.markdown(
+            "<h1 style='text-align: center;'>Login to AppifyAi</h1>",
+            unsafe_allow_html=True,
+        )
+        st.markdown(
+            "<p style='text-align: center;'>Your Personal "
+            "Assistant for Web Application Development</p>",
+            unsafe_allow_html=True,
+        )
 
-        _,c2,_ = st.columns([2,2,2])
+        _, c2, _ = st.columns([2, 2, 2])
+        c2.divider()
 
         form_data, login_message_placeholder = self._create_login_form(c2)
 
@@ -55,36 +67,54 @@ class LoginApp(HydraHeadApp):
         </style>
         <br><br>
         """
-        c2.markdown(pretty_btn,unsafe_allow_html=True)
+        c2.markdown(pretty_btn, unsafe_allow_html=True)
 
-        if form_data['submitted']:
+        if form_data["submitted"]:
             self._do_login(form_data, login_message_placeholder)
 
-
     def _create_login_form(self, parent_container) -> Union[Dict, DeltaGenerator]:
-
         login_form = parent_container.form(key="login_form")
 
         login_message = parent_container.empty()
 
         form_state = {}
-        form_state['username'] = login_form.text_input('Username')
-        form_state['password'] = login_form.text_input('Password',type="password")
-        form_state['submitted'] = login_form.form_submit_button('Login')
+        form_state["username"] = login_form.text_input("Username")
+        form_state["password"] = login_form.text_input("Password", type="password")
+        form_state["submitted"] = login_form.form_submit_button("Login")
 
-        #parent_container.markdown("<p style='text-align: center;'>Guest login -> joe & joe</p>", unsafe_allow_html=True)
+        # Add the javascript to submit the form on enter
+        components.html(
+            """
+        <script>
+        const doc = window.parent.document;
+        buttons = Array.from(doc.querySelectorAll('button[kind=secondaryFormSubmit]'));
+        const submit = buttons.find(el => el.innerText === 'Login');
 
-        if parent_container.button('Sign Up',key='signupbtn'):
-            # set access level to a negative number to allow a kick to the unsecure_app set in the parent
-            self.set_access(-1, 'guest')
+        doc.addEventListener('keydown', function(e) {
+            switch (e.keyCode) {
+                case 13: // (37 = enter)
+                    submit.click();
+            }
+        });
+        </script>
+        """,
+            height=0,
+            width=0,
+        )
 
-            #Do the kick to the signup app
+        if parent_container.button("Sign Up", key="signupbtn"):
+            self.set_access(-1, "guest")
+            self.do_redirect()
+
+        if parent_container.button("What is AppifyAi ?"):
+            self.set_access(1, "guest")
             self.do_redirect()
 
         return form_state, login_message
 
-    def redirect_after_login(self, access_level:int, username:str):
-        #access control uses an int value to allow for levels of permission that can be set for each user, this can then be checked within each app seperately.
+    def redirect_after_login(self, access_level: int, username: str):
+        # access control uses an int value to allow for levels of permission that can be
+        # set for each user, this can then be checked within each app seperately.
         self.set_access(access_level, username, True)
 
         # Seed the sandbox if not already done
@@ -93,9 +123,8 @@ class LoginApp(HydraHeadApp):
         # Init the user data
         self.auth.init_userdata(access_level)
 
-        #Do the kick to the home page
+        # Do the kick to the home page
         self.do_redirect()
-
 
     def _do_login(self, form_data, msg_container) -> None:
         access_level = self._check_login(form_data)
@@ -105,25 +134,31 @@ class LoginApp(HydraHeadApp):
                 self.auth.add_user_session(access_level)
                 with st.spinner("✔️ Login successful, redirecting.."):
                     time.sleep(2)
-                    self.redirect_after_login(access_level, form_data['username'])
+                    self.redirect_after_login(access_level, form_data["username"])
         else:
             self.session_state.allow_access = 0
             self.session_state.current_user = None
             with msg_container:
-                st.error(f"❌ Login unsuccessful, 😕 please check your username and password and try again.")
-
+                st.error(
+                    "❌ Login unsuccessful, 😕 please check your username and password and try again."
+                )
 
     def _check_login(self, login_data) -> int:
-        #this method returns a value indicating the success of verifying the login details provided and the permission level, 1 for default access, 0 no access etc.
-        if self.auth.check_user(login_data['username'], login_data['password']):
-            user_id = AuthSingleton().get_instance().get_user_id(login_data['username'], login_data['password'])
+        # this method returns a value indicating the success of verifying the
+        # login details provided and the permission level, 1 for default access, 0 no access etc.
+        if self.auth.check_user(login_data["username"], login_data["password"]):
+            user_id = (
+                AuthSingleton()
+                .get_instance()
+                .get_user_id(login_data["username"], login_data["password"])
+            )
             return user_id
         return -1
 
     def seed_sandbox(self, level, username):
         # Check if the sandbox exists
-        sandboxes_path = Path(__file__).parent.parent.parent / 'sandboxes'
-        template_sandbox_app = Path(__file__).parent.parent / 'templates' / 'app.py'
+        sandboxes_path = Path(__file__).parent.parent.parent / "sandboxes"
+        template_sandbox_app = Path(__file__).parent.parent / "templates" / "app.py"
         sandbox_user_path = sandboxes_path / f"{username}_{level}.py"
         if sandboxes_path.exists():
             if not sandbox_user_path.exists():
@@ -136,34 +171,49 @@ class LoginApp(HydraHeadApp):
                 try:
                     with st.spinner("Importing sandbox..."):
                         import importlib
-                        spec = importlib.util.spec_from_file_location(f"{username}_{level}", sandbox_user_path)
+
+                        spec = importlib.util.spec_from_file_location(
+                            f"{username}_{level}", sandbox_user_path
+                        )
                         module = importlib.util.module_from_spec(spec)
                         spec.loader.exec_module(module)
+                except BaseException:
+                    self.reset_sandbox(
+                        level, username, template_sandbox_app, sandbox_user_path
+                    )
                 except FileNotFoundError:
-                    with st.spinner("Sandbox needs to be recreated..."):
-                        time.sleep(2)
-                        # Delete the sandbox file
-                        if sandbox_user_path.exists():
-                            os.remove(sandbox_user_path)
-                        # Create the sandbox app
-                        shutil.copyfile(src=template_sandbox_app, dst=sandbox_user_path)
+                    self.reset_sandbox(
+                        level, username, template_sandbox_app, sandbox_user_path
+                    )
 
-                        # Reset chat and code
-                        if "messages" not in st.session_state:
-                            st.session_state["messages"] = []
-                        if "lang" not in st.session_state:
-                            st.session_state["lang"] = "en"
-                        st.session_state.messages = {
-                            "message_0":
-                                {
-                                    "role": "assistant",
-                                    "content": chat_init.message_en.format(name=username) if st.session_state.lang == "en" else chat_init.message_fr.format(name=username)
-                                },
-                        }
-                        if "chat_history" in st.session_state:
-                            st.session_state.chat_history = []
-                        self.auth.set_message_history(level,  st.session_state.messages)
-                        self.auth.set_code(level, "pass")
+        def reset_sandbox(
+            self, level, username, template_sandbox_app, sandbox_user_path
+        ):
+            with st.spinner("Sandbox needs to be recreated..."):
+                time.sleep(2)
+                # Delete the sandbox file
+                if sandbox_user_path.exists():
+                    os.remove(sandbox_user_path)
+                # Create the sandbox app
+                shutil.copyfile(src=template_sandbox_app, dst=sandbox_user_path)
+
+                # Reset chat and code
+                if "messages" not in st.session_state:
+                    st.session_state["messages"] = []
+                if "lang" not in st.session_state:
+                    st.session_state["lang"] = "en"
+                st.session_state.messages = {
+                    "message_0": {
+                        "role": "assistant",
+                        "content": chat_init.message_en.format(name=username)
+                        if st.session_state.lang == "en"
+                        else chat_init.message_fr.format(name=username),
+                    },
+                }
+                if "chat_history" in st.session_state:
+                    st.session_state.chat_history = []
+                self.auth.set_message_history(level, st.session_state.messages)
+                self.auth.set_code(level, "pass")
 
     def reset_chat(self):
         print("resetting chat")