+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-07-19
+
+Change ChatbotX name to AppifyAi
+
+### Added
+
+- demo on face detection
+- add pylint workflow
+- Catch phrase at first
+- secrets for tries
+- Now you can quickly login and signup when enter is pressed
+- About page
+
+### Fixed
+
+- Sandbox was not recreated after a syntax error
+- Code history to undo
+
 ## [1.3.0] - 2023-07-19
 
 ### Added
@@ -47,8 +65,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The chatbot is now able to generate the code and apply it instantly
 - The streamlit generated app is now downloadable
 
-[unreleased]: https://github.com/Gamma-Software/ChatbotX/compare/v1.3.0...HEAD
-[1.3.0]: https://github.com/Gamma-Software/ChatbotX/compare/v1.2.0...v1.3.0
-[1.2.0]: https://github.com/Gamma-Software/ChatbotX/compare/v1.1.0...v1.2.0
-[1.1.0]: https://github.com/Gamma-Software/ChatbotX/compare/v1.0.0...v1.1.0
-[1.0.0]: https://github.com/Gamma-Software/ChatbotX/releases/tag/v1.0.0
+[unreleased]: https://github.com/Gamma-Software/AppifyAi/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/Gamma-Software/AppifyAi/compare/v1.2.0...v1.3.0
+[1.2.0]: https://github.com/Gamma-Software/AppifyAi/compare/v1.1.0...v1.2.0
+[1.1.0]: https://github.com/Gamma-Software/AppifyAi/compare/v1.0.0...v1.1.0
+[1.0.0]: https://github.com/Gamma-Software/AppifyAi/releases/tag/v1.0.0